Compiling `sd_initer.c`.
