+++ electrical/sd_initer.c
@@ -1,3 +1,87 @@
+#include "sd.meta"
+
+
+
+////////////////////////////////////////////////////////////////////////////////
+
+enum SDIniterErr : u32
+{
+    SDIniterErr_none,
+    SDIniterErr_bug,
+    SDIniterErr_sorry,
+};
+
+
+
+enum SDCardState : u32 // @/pg 148/tbl 4-42/`SD`.
+{
+    SDCardState_idle  = 0,
+    SDCardState_ready = 1,
+    SDCardState_ident = 2, // "Identification".
+    SDCardState_stby  = 3, // "Stand-by".
+    SDCardState_tran  = 4, // "Transfer".
+    SDCardState_data  = 5,
+    SDCardState_rcv   = 6, // "Receive-data".
+    SDCardState_prg   = 7, // "Programming".
+    SDCardState_dis   = 8, // "Disconnect".
+};
+
+
+
+enum SDIniterState : u32
+{
+    SDIniterState_uninited,
+    SDIniterState_GO_IDLE_STATE,
+    SDIniterState_SEND_IF_COND,
+    SDIniterState_SD_SEND_OP_COND,
+    SDIniterState_ALL_SEND_CID,
+    SDIniterState_SEND_RELATIVE_ADDR,
+    SDIniterState_SEND_CSD,
+    SDIniterState_SELECT_DESELECT_CARD,
+    SDIniterState_SEND_SCR,
+    SDIniterState_SWITCH_FUNC,
+    SDIniterState_SET_BUS_WIDTH,
+    SDIniterState_caller_set_bus_width,
+    SDIniterState_done,
+};
+
+
+
+struct SDIniter
+{
+    enum SDIniterState state;
+    u16                rca; // Relative card address. @/pg 67/sec 4.2.2/`SD`.
+    i32                capacity_sector_count;
+
+    struct
+    {
+        i32 mount_attempts;
+        i32 support_attempts;
+        u8  sd_config_reg[8];
+        u8  switch_func_status[64];
+    } local;
+
+    struct SDIniterFeedback // To be filled out by the caller before updating with the results of the previous command.
+    {
+        b32 cmd_failed;
+        u32 response[4]; // Card command response with the first word being the most-significant bits of the response.
+    } feedback;
+
+    struct SDIniterCmd
+    {
+        enum SDCmd cmd;
+        u32        arg;
+        u8*        data;
+        i32        size;
+    } cmd;
+};
+
+
+
+////////////////////////////////////////////////////////////////////////////////
+
+
+
 static useret enum SDIniterHandleFeedbackRet : u32
 {
     SDIniterHandleFeedbackRet_ok,
@@ -7,10 +91,9 @@ static useret enum SDIniterHandleFeedbackRet : u32
 }
 _SD_initer_handle_feedback(struct SDIniter* initer)
 {
-    #undef  ret
-    #define ret(KIND) ret_mk(SDIniterHandleFeedbackRet, KIND)
 
-    ret_if(bug, !initer);
+    if (!initer)
+        return SDIniterHandleFeedbackRet_bug;
 
     initer->cmd = (struct SDIniterCmd) {0};
 
@@ -23,7 +106,8 @@ _SD_initer_handle_feedback(struct SDIniter* initer)
         case SDIniterState_uninited:
         {
 
-            ret_if(bug, initer->feedback.cmd_failed); // Feedback said SD command failed, but SD-initer hasn't given the caller any SD commands yet.
+            if (initer->feedback.cmd_failed)
+                return SDIniterHandleFeedbackRet_bug; // Feedback said SD command failed, but SD-initer hasn't given the caller any SD commands yet.
 
             initer->state = SDIniterState_GO_IDLE_STATE;
 
@@ -36,7 +120,8 @@ _SD_initer_handle_feedback(struct SDIniter* initer)
         case SDIniterState_GO_IDLE_STATE:
         {
 
-            ret_if(support_issue, initer->feedback.cmd_failed);
+            if (initer->feedback.cmd_failed)
+                return SDIniterHandleFeedbackRet_support_issue;
 
             initer->state = SDIniterState_SEND_IF_COND;
 
@@ -58,7 +143,8 @@ _SD_initer_handle_feedback(struct SDIniter* initer)
 
             if (initer->feedback.cmd_failed)
             {
-                ret_if(card_likely_unmounted, initer->local.mount_attempts + 1 >= 64);
+                if (initer->local.mount_attempts + 1 >= 64)
+                    return SDIniterHandleFeedbackRet_card_likely_unmounted;
 
                 initer->local.mount_attempts += 1;
                 initer->state                 = SDIniterState_GO_IDLE_STATE; // Try again.
@@ -66,7 +152,8 @@ _SD_initer_handle_feedback(struct SDIniter* initer)
             else
             {
                 // Response of SEND_IF_COND must echo back the expected voltage and check-pattern. @/pg 116/sec 4.3.13/`SD`."
-                ret_if(support_issue, initer->feedback.response[0] != (SD_INTERFACE_CONDITION & ((1 << 12) - 1)));
+                if (initer->feedback.response[0] != (SD_INTERFACE_CONDITION & ((1 << 12) - 1)))
+                    return SDIniterHandleFeedbackRet_support_issue;
 
                 initer->local.mount_attempts = 0;
                 initer->state                = SDIniterState_SD_SEND_OP_COND;
@@ -86,7 +173,8 @@ _SD_initer_handle_feedback(struct SDIniter* initer)
 
             if (initer->feedback.cmd_failed || !powered_up) // Couldn't ready the card?
             {
-                ret_if(card_likely_unmounted, initer->local.mount_attempts + 1 >= 1024);
+                if (initer->local.mount_attempts + 1 >= 1024)
+                    return SDIniterHandleFeedbackRet_card_likely_unmounted;
 
                 initer->local.mount_attempts += 1;
                 initer->state                 = SDIniterState_SD_SEND_OP_COND; // Try again.
@@ -95,7 +183,8 @@ _SD_initer_handle_feedback(struct SDIniter* initer)
             {
                 // Check the CCS bit where "CCS=0 is SDSC and CCS=1 is SDHC or SDXC". @/pg 117/sec 4.3.14/`SD`.
                 // We don't support SDSC; if we were to, we'd need to account for the byte unit addressing. @/pg 130/tbl 4-23/`SD`.
-                ret_if(support_issue, !(ocr & (1 << 30)));
+                if (!(ocr & (1 << 30)))
+                    return SDIniterHandleFeedbackRet_support_issue;
 
                 initer->local.mount_attempts = 0;
                 initer->state                = SDIniterState_ALL_SEND_CID;
@@ -110,7 +199,8 @@ _SD_initer_handle_feedback(struct SDIniter* initer)
         case SDIniterState_ALL_SEND_CID:
         {
 
-            ret_if(support_issue, initer->feedback.cmd_failed);
+            if (initer->feedback.cmd_failed)
+                return SDIniterHandleFeedbackRet_support_issue;
 
             initer->state = SDIniterState_SEND_RELATIVE_ADDR; // The card is now in the identification state.
 
@@ -123,8 +213,10 @@ _SD_initer_handle_feedback(struct SDIniter* initer)
         case SDIniterState_SEND_RELATIVE_ADDR:
         {
 
-            ret_if(support_issue, initer->feedback.cmd_failed);
-            ret_if(support_issue, ((initer->feedback.response[0] >> 9) & 0b1111) != SDCardState_ident); // Transitioned out of the identification state.
+            if (initer->feedback.cmd_failed)
+                return SDIniterHandleFeedbackRet_support_issue;
+            if (((initer->feedback.response[0] >> 9) & 0b1111) != SDCardState_ident)
+                return SDIniterHandleFeedbackRet_support_issue; // Transitioned out of the identification state.
 
             initer->rca   = initer->feedback.response[0] >> 16; // The RCA is needed for some commands. @/pg 144/sec 4.9.5/`SD`.
             initer->state = SDIniterState_SEND_CSD;
@@ -138,7 +230,8 @@ _SD_initer_handle_feedback(struct SDIniter* initer)
         case SDIniterState_SEND_CSD:
         {
 
-            ret_if(support_issue, initer->feedback.cmd_failed);
+            if (initer->feedback.cmd_failed)
+                return SDIniterHandleFeedbackRet_support_issue;
 
             union
             {
@@ -152,20 +245,22 @@ _SD_initer_handle_feedback(struct SDIniter* initer)
             }};
 
             struct SDCardSpecificData csd = {0};
-            ret_if(support_issue, !SD_parse_register(&csd, card_specific_data_data.bytes));
+            if (!SD_parse_register(&csd, card_specific_data_data.bytes))
+                return SDIniterHandleFeedbackRet_support_issue;
 
             switch (csd.CSD_STRUCTURE)
             {
                 case SDCardSpecificDataVersion_2:
                 {
                     // Classes 0, 2, 4, 5, and 8 are mandatory. @/pg 123/sec 4.7.3/`SD`.
-                    ret_if(support_issue, ~csd.v2_CCC & ((1 << 0) | (1 << 2) | (1 << 4) | (1 << 5) | (1 << 8)));
+                    if (~csd.v2_CCC & ((1 << 0) | (1 << 2) | (1 << 4) | (1 << 5) | (1 << 8)))
+                        return SDIniterHandleFeedbackRet_support_issue;
 
                     // Memory capacity of the SD card in units of 512-byte sectors. @/pg 260/sec 5.3.3/`SD`.
                     initer->capacity_sector_count = (csd.v2_C_SIZE + 1) * 1024;
                 } break;
 
-                default: ret(support_issue);
+                default: return SDIniterHandleFeedbackRet_support_issue;
             }
 
             initer->state = SDIniterState_SELECT_DESELECT_CARD;
@@ -179,8 +274,10 @@ _SD_initer_handle_feedback(struct SDIniter* initer)
         case SDIniterState_SELECT_DESELECT_CARD:
         {
 
-            ret_if(support_issue, initer->feedback.cmd_failed);
-            ret_if(support_issue, ((initer->feedback.response[0] >> 9) & 0b1111) != SDCardState_stby); // Transitioned out of the stand-by state. @/pg 76/fig 4-13/`SD`.
+            if (initer->feedback.cmd_failed)
+                return SDIniterHandleFeedbackRet_support_issue;
+            if (((initer->feedback.response[0] >> 9) & 0b1111) != SDCardState_stby)
+                return SDIniterHandleFeedbackRet_support_issue; // Transitioned out of the stand-by state. @/pg 76/fig 4-13/`SD`.
 
             initer->state = SDIniterState_SEND_SCR;
 
@@ -195,22 +292,27 @@ _SD_initer_handle_feedback(struct SDIniter* initer)
 
             struct SDConfigurationRegister scr = {0};
 
-            ret_if(support_issue, initer->feedback.cmd_failed);
-            ret_if(support_issue, ((initer->feedback.response[0] >> 9) & 0b1111) != SDCardState_tran); // Transitioned out of the transmission state. @/pg 76/fig 4-13/`SD`.
-            ret_if(support_issue, !SD_parse_register(&scr, initer->local.sd_config_reg));
+            if (initer->feedback.cmd_failed)
+                return SDIniterHandleFeedbackRet_support_issue;
+            if (((initer->feedback.response[0] >> 9) & 0b1111) != SDCardState_tran)
+                return SDIniterHandleFeedbackRet_support_issue; // Transitioned out of the transmission state. @/pg 76/fig 4-13/`SD`.
+            if (!SD_parse_register(&scr, initer->local.sd_config_reg))
+                return SDIniterHandleFeedbackRet_support_issue;
 
             switch (scr.SCR_STRUCTURE)
             {
                 case SDConfigurationRegisterVersion_1:
                 {
                     // Only SD cards with version 2.00 of the physical layer specification are currently supported. @/pg 267/tbl 5-19/`SD`
-                    ret_if(support_issue, scr.v1_SD_SPEC != 2);
+                    if (scr.v1_SD_SPEC != 2)
+                        return SDIniterHandleFeedbackRet_support_issue;
 
                     // Only SD cards with 1-bit and 4-bit busses are currently supported. @/pg 270/tbl 5-21/`SD`.
-                    ret_if(support_issue, ~scr.v1_SD_BUS_WIDTHS & ((1 << 0) | (1 << 2)));
+                    if (~scr.v1_SD_BUS_WIDTHS & ((1 << 0) | (1 << 2)))
+                        return SDIniterHandleFeedbackRet_support_issue;
                 } break;
 
-                default: ret(support_issue);
+                default: return SDIniterHandleFeedbackRet_support_issue;
             }
 
             initer->state = SDIniterState_SWITCH_FUNC;
@@ -226,8 +328,10 @@ _SD_initer_handle_feedback(struct SDIniter* initer)
 
             struct SDSwitchFunctionStatus sfs = {0};
 
-            ret_if(support_issue, initer->feedback.cmd_failed);
-            ret_if(support_issue, !SD_parse_register(&sfs, initer->local.switch_func_status));
+            if (initer->feedback.cmd_failed)
+                return SDIniterHandleFeedbackRet_support_issue;
+            if (!SD_parse_register(&sfs, initer->local.switch_func_status))
+                return SDIniterHandleFeedbackRet_support_issue;
 
             switch (sfs.version)
             {
@@ -235,13 +339,15 @@ _SD_initer_handle_feedback(struct SDIniter* initer)
                 case SDSwitchFunctionStatusVersion_2:
                 {
                     // Only SD cards with high-speed mode are currently supported. @/pg 111/tbl 4-13/`SD`.
-                    ret_if(support_issue, ~sfs.v2_group_1_support & (1 << 1));
+                    if (~sfs.v2_group_1_support & (1 << 1))
+                        return SDIniterHandleFeedbackRet_support_issue;
 
                     // SD cards couldn't be switched to high-speed mode. @/pg 107/tbl 4-11/`SD`.
-                    ret_if(support_issue, sfs.v2_group_1_selection != 0x1);
+                    if (sfs.v2_group_1_selection != 0x1)
+                        return SDIniterHandleFeedbackRet_support_issue;
                 } break;
 
-                default: ret(support_issue);
+                default: return SDIniterHandleFeedbackRet_support_issue;
             }
 
             initer->state = SDIniterState_SET_BUS_WIDTH;
@@ -255,8 +361,10 @@ _SD_initer_handle_feedback(struct SDIniter* initer)
         case SDIniterState_SET_BUS_WIDTH:
         {
 
-            ret_if(support_issue, initer->feedback.cmd_failed);
-            ret_if(support_issue, ((initer->feedback.response[0] >> 9) & 0b1111) != SDCardState_tran); // Transitioned out of the transfer state. @/pg 76/fig 4-13/`SD`.
+            if (initer->feedback.cmd_failed)
+                return SDIniterHandleFeedbackRet_support_issue;
+            if (((initer->feedback.response[0] >> 9) & 0b1111) != SDCardState_tran)
+                return SDIniterHandleFeedbackRet_support_issue; // Transitioned out of the transfer state. @/pg 76/fig 4-13/`SD`.
 
             initer->state = SDIniterState_caller_set_bus_width;
 
@@ -269,18 +377,21 @@ _SD_initer_handle_feedback(struct SDIniter* initer)
         case SDIniterState_caller_set_bus_width:
         {
 
-            ret_if(bug, initer->feedback.cmd_failed); // No reason for the caller to fail configuring the bus width.
+            if (initer->feedback.cmd_failed)
+                return SDIniterHandleFeedbackRet_bug; // No reason for the caller to fail configuring the bus width.
 
             initer->state = SDIniterState_done;
 
         } break;
 
-        case SDIniterState_done : ret(bug); // We're already done; there's nothing left to do!
-        default                 : ret(bug);
+        case SDIniterState_done : return SDIniterHandleFeedbackRet_bug; // We're already done; there's nothing left to do!
+        default                 : return SDIniterHandleFeedbackRet_bug;
     }
 
-    ret(ok);
+    return SDIniterHandleFeedbackRet_ok;
 }
+
+
 
 static useret enum SDIniterUpdateRet : u32
 {
@@ -293,10 +404,9 @@ static useret enum SDIniterUpdateRet : u32
 }
 _SD_initer_update(struct SDIniter* initer)
 {
-    #undef  ret
-    #define ret(KIND) ret_mk(SDIniterUpdateRet, KIND)
 
-    ret_if(bug, !initer);
+    if (!initer)
+        return SDIniterUpdateRet_bug;
 
     //
     // Handle feedback.
@@ -307,16 +417,17 @@ _SD_initer_update(struct SDIniter* initer)
     {
         case SDIniterHandleFeedbackRet_support_issue:
         {
-            ret_if(card_likely_unsupported, initer->local.support_attempts + 1 >= 64);
+            if (initer->local.support_attempts + 1 >= 64)
+                return SDIniterUpdateRet_card_likely_unsupported;
 
             initer->local.support_attempts += 1;
             initer->state                   = SDIniterState_GO_IDLE_STATE; // Redo the initialization again.
         } break;
 
         case SDIniterHandleFeedbackRet_ok                    : break;
-        case SDIniterHandleFeedbackRet_card_likely_unmounted : ret(card_likely_unmounted);
-        case SDIniterHandleFeedbackRet_bug                   : ret(bug);
-        default                                              : ret(bug);
+        case SDIniterHandleFeedbackRet_card_likely_unmounted : return SDIniterUpdateRet_card_likely_unmounted;
+        case SDIniterHandleFeedbackRet_bug                   : return SDIniterUpdateRet_bug;
+        default                                              : return SDIniterUpdateRet_bug;
     }
 
     //
@@ -330,7 +441,7 @@ _SD_initer_update(struct SDIniter* initer)
             do                                                      \
             {                                                       \
                 initer->cmd = (struct SDIniterCmd) { __VA_ARGS__ }; \
-                ret(do_cmd);                                        \
+                return SDIniterUpdateRet_do_cmd;                    \
             }                                                       \
             while (false)
 
@@ -378,9 +489,9 @@ _SD_initer_update(struct SDIniter* initer)
                 .size = sizeof(initer->local.switch_func_status),
             );
 
-        case SDIniterState_caller_set_bus_width : ret(caller_set_bus_width);
-        case SDIniterState_done                 : ret(done);
-        case SDIniterState_uninited             : ret(bug);
-        default                                 : ret(bug);
+        case SDIniterState_caller_set_bus_width : return SDIniterUpdateRet_caller_set_bus_width;
+        case SDIniterState_done                 : return SDIniterUpdateRet_done;
+        case SDIniterState_uninited             : return SDIniterUpdateRet_bug;
+        default                                 : return SDIniterUpdateRet_bug;
     }
 }
